Add files via upload
This notebook implements an optimal control model to study and manage the spread of a disease. The model is based on a system of ordinary differential equations (ODEs) that describe the population dynamics in different compartments:

S: Susceptible individuals
P: Protected individuals (e.g., through vaccination or maternal immunity)
C_A: Asymptomatic carriers
I_A: Infectious individuals
R: Recovered individuals
The objective of the model is to find the optimal control strategies, specifically vaccination and treatment rates, that minimise the number of infectious individuals over a given time horizon while considering the cost of implementing these controls.

The code uses scipy.integrate.The solve_ivp function is used to solve the ODE system and scipy. optimise. minimise to find the optimal control functions by minimising a defined cost function. The cost function penalises both the number of infectious individuals and the magnitude of the control efforts (vaccination and treatment rates).

The results of the optimisation and simulation are visualised through plots showing the population dynamics in each compartment and the time evolution of the optimal vaccination and treatment rates.